# app/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about\
    \     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net,\
    \ #swagger](http://swagger.io/irc/).      For this sample, you can use the api\
    \ key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "FHW Calendar Exporter"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v2"
tags:
- name: "splan"
  description: "Everything about the splan exporter"
  externalDocs:
    description: "Find out more"
    url: "https://intern.fh-wedel.de/~splan/"
schemes:
- "https"
- "http"
paths:
  /splan/json:
    get:
      tags:
      - "splan"
      summary: "Export the splan to JSON"
      operationId: "getSplanAsJson"
      produces:
      - "application/json"
      parameters:
      - name: "url"
        in: "query"
        description: "URL of custom splan"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Plan"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Splan"
  /splan/ics:
    get:
      tags:
      - "splan"
      summary: "Export the splan to the ICalender format"
      operationId: "getSplanAsIcs"
      produces:
      - "text/calendar"
      parameters:
      - name: "url"
        in: "query"
        description: "URL of custom splan"
        required: true
        type: "string"
      - name: "nameformat"
        in: "query"
        description: "Format of person names in output"
        required: false
        type: "string"
        default: "abbr"
        enum:
        - "none"
        - "abbr"
        - "first_last"
        - "first_last_abbr"
        - "last_first"
        - "last_first_abbr"
      - name: "roomformat"
        in: "query"
        description: "Format of room names in output"
        required: false
        type: "string"
        default: "abbr"
        enum:
        - "none"
        - "abbr"
        - "first_last"
        - "first_last_abbr"
        - "last_first"
        - "last_first_abbr"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Splan"
  /splan/pdf:
    get:
      tags:
      - "splan"
      summary: "Export the splan to PDF"
      operationId: "getSplanAsPdf"
      produces:
      - "application/pdf"
      parameters:
      - name: "url"
        in: "query"
        description: "URL of custom splan"
        required: true
        type: "string"
      - name: "nameformat"
        in: "query"
        description: "Format of person names in output"
        required: false
        type: "string"
        default: "abbr"
        enum:
        - "none"
        - "abbr"
        - "first_last"
        - "first_last_abbr"
        - "last_first"
        - "last_first_abbr"
      - name: "roomformat"
        in: "query"
        description: "Format of room names in output"
        required: false
        type: "string"
        default: "abbr"
        enum:
        - "none"
        - "abbr"
        - "first_last"
        - "first_last_abbr"
        - "last_first"
        - "last_first_abbr"
      - name: "orientation"
        in: "query"
        description: "Orientation of the the table (horizontal or vertical)"
        required: false
        type: "string"
        default: "h"
        enum:
        - "h"
        - "v"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Splan"
definitions:
  Plan:
    type: "object"
    properties:
      title:
        type: "string"
        format: "string"
        example: "B_Inf5"
      courses:
        type: "array"
        items:
          $ref: "#/definitions/Course"
    example:
      courses:
      - rooms:
        - name: "Hörsaal 6"
          abbr: "HS06"
        - name: "Hörsaal 6"
          abbr: "HS06"
        name: "FOOP"
        time:
          start: "8:00"
          length: 1
          end: "9:15"
          day: 1
        employees:
        - last_name: "Uhlig"
          abbr: "uhl"
          first_name: "Christian"
        - last_name: "Uhlig"
          abbr: "uhl"
          first_name: "Christian"
      - rooms:
        - name: "Hörsaal 6"
          abbr: "HS06"
        - name: "Hörsaal 6"
          abbr: "HS06"
        name: "FOOP"
        time:
          start: "8:00"
          length: 1
          end: "9:15"
          day: 1
        employees:
        - last_name: "Uhlig"
          abbr: "uhl"
          first_name: "Christian"
        - last_name: "Uhlig"
          abbr: "uhl"
          first_name: "Christian"
      title: "B_Inf5"
  Course:
    type: "object"
    properties:
      ID:
        type: "string"
        format: "string"
        example: "SB024"
      name:
        type: "string"
        format: "string"
        example: "FOOP"
      time:
        $ref: "#/definitions/TimeInfo"
      employees:
        type: "array"
        items:
          $ref: "#/definitions/Employee"
      rooms:
        type: "array"
        items:
          $ref: "#/definitions/Room"
    example:
      rooms:
      - name: "Hörsaal 6"
        abbr: "HS06"
      - name: "Hörsaal 6"
        abbr: "HS06"
      name: "FOOP"
      time:
        start: "8:00"
        length: 1
        end: "9:15"
        day: 1
      employees:
      - last_name: "Uhlig"
        abbr: "uhl"
        first_name: "Christian"
      - last_name: "Uhlig"
        abbr: "uhl"
        first_name: "Christian"
  TimeInfo:
    type: "object"
    properties:
      day:
        type: "integer"
        format: "int64"
        example: 1
        minimum: 0
        maximum: 6
      start:
        type: "string"
        format: "string"
        example: "8:00"
      end:
        type: "string"
        format: "string"
        example: "9:15"
      length:
        type: "integer"
        format: "int64"
        example: 1
    example:
      start: "8:00"
      length: 1
      end: "9:15"
      day: 1
  Employee:
    type: "object"
    properties:
      first_name:
        type: "string"
        format: "string"
        example: "Christian"
      last_name:
        type: "string"
        format: "string"
        example: "Uhlig"
      abbr:
        type: "string"
        format: "string"
        example: "uhl"
    example:
      last_name: "Uhlig"
      abbr: "uhl"
      first_name: "Christian"
  Room:
    type: "object"
    properties:
      name:
        type: "string"
        format: "string"
        example: "Hörsaal 6"
      abbr:
        type: "string"
        format: "string"
        example: "HS06"
    example:
      name: "Hörsaal 6"
      abbr: "HS06"
  NameFormat:
    type: "string"
    enum:
    - "none"
    - "abbr"
    - "first_last"
    - "first_last_abbr"
    - "last_first"
    - "last_first_abbr"
  RoomFormat:
    type: "string"
    enum:
    - "none"
    - "abbr"
    - "first_last"
    - "first_last_abbr"
    - "last_first"
    - "last_first_abbr"
parameters:
  urlParam:
    name: "url"
    in: "query"
    description: "URL of custom splan"
    required: true
    type: "string"
  nameFormatParam:
    name: "nameformat"
    in: "query"
    description: "Format of person names in output"
    required: false
    type: "string"
    default: "abbr"
    enum:
    - "none"
    - "abbr"
    - "first_last"
    - "first_last_abbr"
    - "last_first"
    - "last_first_abbr"
  roomFormatParam:
    name: "roomformat"
    in: "query"
    description: "Format of room names in output"
    required: false
    type: "string"
    default: "abbr"
    enum:
    - "none"
    - "abbr"
    - "first_last"
    - "first_last_abbr"
    - "last_first"
    - "last_first_abbr"
  orientationParam:
    name: "orientation"
    in: "query"
    description: "Orientation of the the table (horizontal or vertical)"
    required: false
    type: "string"
    default: "h"
    enum:
    - "h"
    - "v"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"